chore(northbrook): start from nothing :smile:
Too many pain points have been felt when actually using
Northbrook for my projects, and I genuinely feel there
needs to be a simplifying of the concepts. I've begun to feel
as though Northbrook is attempting to live a too high a level of
abstraction and it's scope needs to narrow and become largely
more opionated about how it manages and works with plugins, potentially
defining one or more functionalities that **are** default and can not
be changed in order to facilate better practices for open-source
projects.